pillar.environments: handle possible RecursionError
seems there are some latent race condition bugs when a machine exists in
bbcpu but not bbmachines/cluster.json. this stops the RecursionError but
this means we still have situations where we dont have the right data
which could impact newly built machines. implementing the sor fallback
is probably the correct course of action to fix that.