auth_check: ensure recursive marked acl take priority
because external_auth is a merged list of acl rules, they are unordered.
with the introduction of recursive, a scenario could occur where a user
could have permissions to run something with both recursive and
non-recursive perms, and ordering is based off config. this ensures
recursive will always take priority over non recursive.