re-work auth_check authentication
we need to verify user/key based auth regardless of auth_check coming in
off the wire. we also need to respect recursive if it is coming from an
authenticated request and not re-authorize.

re-work auth_check authentication, take 2

This reverts commit 4b8867f9facfe7aab1b8f1fa1598666bce280cbd.

passing recursive or auth_check via clear load is not something we can
trust to be safe. instead we relay token/eauth credentials on each
nested call, and re-authenticate on each entry.