saltnado: change to tornado.queues per request
the previous implementation used future per event with a large amount of
shuffling to make this work for localclient requests. there are also
various bugs that have been associated with it (#35798, probably others
mim too lazy to find). This changes each EventListener instance to map
to a queue/request/tag tuple that on_recv populates as events come in.
This ensures events aren't consumed that should not be, and various
other weirdness. I also gutted how job_not_running as it wasn't doing
what it said on the tin.