master.publish/minion acl: handle recursive
we weren't pushing down the auth_check in publish load (I accidentally
left it in with the opts_override patchset). This handles recursive by
_not including it at all_ if the bit is set, which should function
equivalently to a user/cli job that has no acl at all. Once the fleet
has been converted we can switch the commented if with the other to
allow auth_check propogation from master to minion.