new feature: zmq_filtering discovery by way of master
rather than rely on a non-default minion opt to use in conjunction with
enabling zmq_filtering master side, it is infinitely more effective to
let each master tell its minion which mode it is operating in.
This should allow a multi-master minion to connect to mixed set of masters
where some are enabled for zmq_filtering, and some are not.